requirements.txt
anyio==4.3.0
arq==0.25.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
corpwechatbot==0.6.4
cptools==2.0.1
cryptography==42.0.5
Django==5.0.2
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-monthfield==0.1.4
django-polymorphic==3.1.0
django-subcommands==2020.12.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
fire==0.5.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
json-fix==1.0.0
mysqlclient==2.2.4
numpy==2.0.0
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
pika==1.3.2
playwright==1.41.2
pluggy==1.4.0
pycparser==2.21
pyee==11.0.1
PyMySQL==1.1.0
pytest==8.0.2
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sqltree==0.3.0
stringcase==1.2.0
termcolor==2.4.0
text-unidecode==1.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
